+++ monolith/src/main/java/com/skripsi/monolith/repository/country/CountryRepository.java
@@ -1,6 +1,6 @@
-package com.skripsi.monolith.repository;
+package com.skripsi.monolith.repository.country;
 
-import com.skripsi.monolith.model.Country;
+import com.skripsi.monolith.model.country.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
